Refactor: 파일 경로 이동 및 SideBar 스타일 파일 분리
