+++ src/widgets/common/Footer.tsx
