+++ src/widgets/common/WidgetHeader.tsx
